Fix typo in kernel and USE THE RIGHT VALUE FOR PI

+++ Hydro-master/mathdef.h
@@ -1,7 +1,7 @@
 #ifndef _MATHDEF_H_
 #define _MATHDEF_H_
 
-#define PI (4*atan((float)1))
+#define PI (4*atan((long double)1))
 
 #define sqr(x) ((x)*(x))
